+++ Modules/Ruleset.ProgramInfo/Public/Get-SystemApp.ps1
@@ -154,6 +154,9 @@ function Get-SystemApp
 		Write-Verbose -Message "[$($MyInvocation.InvocationName)] No apps were found for '$User' user on '$Domain'"
 	}
 
+	# Index 0 is this function
+	$Caller = (Get-PSCallStack)[1].Command
+
 	foreach ($App in $Apps)
 	{
 		if (($App.SignatureKind -eq "System") -and ($App.Name -like "Microsoft*"))
@@ -173,7 +176,7 @@ function Get-SystemApp
 			{
 				Write-Warning -Message "[$($MyInvocation.InvocationName)] Store app '$($App.Name)' is not installed by user '$User' or the app is missing"
 				Write-Information -Tags $MyInvocation.InvocationName `
-					-MessageData "INFO: To fix the problem let this user update all of it's apps in Windows store"
+					-MessageData "INFO: To fix this problem let this user update all of it's apps in Windows store, then rerun '$Caller' script"
 			}
 		}
 	}