+++ Modules/Ruleset.ProgramInfo/Public/Get-UserApp.ps1
@@ -160,6 +160,9 @@ function Get-UserApp
 		Write-Verbose -Message "[$($MyInvocation.InvocationName)] No apps were found for '$User' user on '$Domain'"
 	}
 
+	# Index 0 is this function
+	$Caller = (Get-PSCallStack)[1].Command
+
 	foreach ($App in $Apps)
 	{
 		# NOTE: This path will be missing for default apps on Windows server
@@ -177,7 +180,7 @@ function Get-UserApp
 		{
 			Write-Warning -Message "[$($MyInvocation.InvocationName)] Store app '$($App.Name)' is not installed by user '$User' or the app is missing"
 			Write-Information -Tags $MyInvocation.InvocationName `
-				-MessageData "INFO: To fix the problem let this user update all of it's apps in Windows store"
+				-MessageData "INFO: To fix the problem let this user update all of it's apps in Windows store, then rerun '$Caller' script"
 		}
 	}
 }